Move utils to its file
